refactor: move login controller to view model
